requirements.txt
streamlit==1.28.0
Pillow==10.1.0
torch==2.2.2
transformers==4.38.2
sentence-transformers==2.5.1
openai==1.16.2
langchain==0.2.1
langchain-community==0.0.30
langchain-huggingface==0.0.21
langchain-openai==0.1.14
biopython==1.83
